bug fix, refactoring
fixed setDelay bug.
disabled hazelcast, akka info logs.
moved DTOs into separate module.
added lib folder for jar files.
added DTO jar as dependency to TGF
