+++ TGF-CHF-DTO/src/main/java/org/sk/i2i/evren/DataTransaction.java
@@ -1,4 +1,4 @@
-package org.sk.i2i.evren.TGF.DTO;
+package org.sk.i2i.evren;
 
 public class DataTransaction {
