+++ TGF-CHF-DTO/src/main/java/org/sk/i2i/evren/SmsTransaction.java
@@ -1,4 +1,4 @@
-package org.sk.i2i.evren.TGF.DTO;
+package org.sk.i2i.evren;
 
 public class SmsTransaction {
     private final String senderMsisdn;
