+++ TGF/pom.xml
@@ -61,6 +61,14 @@
         </dependency>
 
         <dependency>
+            <groupId>org.sk.i2i.evren</groupId>
+            <artifactId>TGF-CHF-DTO</artifactId>
+            <version>1.0-SNAPSHOT</version>
+            <scope>system</scope>
+            <systemPath>${project.basedir}/src/main/resources/TGF-CHF-DTO.jar</systemPath>
+        </dependency>
+
+        <dependency>
             <groupId>com.hazelcast</groupId>
             <artifactId>hazelcast</artifactId>
             <version>5.4.0</version>
