+++ TGF/src/main/java/org/sk/i2i/evren/TGF/actors/AkkaActor.java
@@ -3,9 +3,9 @@ package org.sk.i2i.evren.TGF.actors;
 import akka.actor.AbstractActor;
 import akka.actor.ActorSelection;
 import com.typesafe.config.Config;
-import org.sk.i2i.evren.TGF.DTO.DataTransaction;
-import org.sk.i2i.evren.TGF.DTO.SmsTransaction;
-import org.sk.i2i.evren.TGF.DTO.VoiceTransaction;
+import org.sk.i2i.evren.DataTransaction;
+import org.sk.i2i.evren.SmsTransaction;
+import org.sk.i2i.evren.VoiceTransaction;
 
 public class AkkaActor extends AbstractActor {
 
@@ -15,7 +15,7 @@ public class AkkaActor extends AbstractActor {
     @Override
     public Receive createReceive() {
         return receiveBuilder()
-                .match( DataTransaction.class, trans -> remoteActor.tell(trans, getSelf()) )
+                .match(DataTransaction.class, trans -> remoteActor.tell(trans, getSelf()) )
                 .match( VoiceTransaction.class, trans -> remoteActor.tell(trans, getSelf()) )
                 .match( SmsTransaction.class, trans -> remoteActor.tell(trans, getSelf()) )
                 .matchAny(o -> System.out.println("received unknown message"))
