+++ TGF/src/main/java/org/sk/i2i/evren/TGF/actors/DeadLetterListener.java
@@ -2,10 +2,10 @@ package org.sk.i2i.evren.TGF.actors;
 
 import akka.actor.AbstractActor;
 import akka.actor.DeadLetter;
-import org.sk.i2i.evren.TGF.DTO.DataTransaction;
-import org.sk.i2i.evren.TGF.DTO.SmsTransaction;
-import org.sk.i2i.evren.TGF.DTO.VoiceTransaction;
+import org.sk.i2i.evren.DataTransaction;
+import org.sk.i2i.evren.SmsTransaction;
 import org.sk.i2i.evren.TGF.DeadLetterStats;
+import org.sk.i2i.evren.VoiceTransaction;
 
 public class DeadLetterListener extends AbstractActor {
     DeadLetterStats deadLetterStats;
