+++ TGF/src/main/resources/application.conf
@@ -6,9 +6,9 @@ akka {
           jackson-json = "akka.serialization.jackson.JacksonJsonSerializer"
         }
         serialization-bindings {
-          "org.sk.i2i.evren.TGF.DTO.DataTransaction" = jackson-json
-          "org.sk.i2i.evren.TGF.DTO.VoiceTransaction" = jackson-json
-          "org.sk.i2i.evren.TGF.DTO.SmsTransaction" = jackson-json
+          "org.sk.i2i.evren.DataTransaction" = jackson-json
+          "org.sk.i2i.evren.VoiceTransaction" = jackson-json
+          "org.sk.i2i.evren.SmsTransaction" = jackson-json
         }
   }
   remote {
@@ -17,6 +17,7 @@ akka {
       canonical.port = 25520
     }
   }
+  loglevel = "ERROR"
 }
 
 CHF.path = "akka://TGFSystemCopy@127.0.0.1:25521/user/TGFActorCopy"
